feat: update linking system to account for documentation version
