+++ src/components/Markdown/Markdown.tsx
@@ -1,12 +1,15 @@
 'use client'
 
-import { Marked, Renderer } from '@ts-stack/markdown'
+import { Marked, MarkedOptions, Renderer } from '@ts-stack/markdown'
 import style from './Markdown.module.css'
 import hljs, { LanguageFn } from 'highlight.js'
 import amber from './amber'
-import { useEffect } from 'react'
+import { useEffect, useMemo } from 'react'
 import setSwipeToCopy from './swipeToCopy'
 import complexImageParser, { COMPLEX_IMAGE_RULE } from './complexImage'
+import useVersion from '@/contexts/VersionContext/useVersion'
+import { generateUrl } from '@/utils/urls'
+import { MarkOptions } from 'perf_hooks'
 
 hljs.registerLanguage('amber', amber as LanguageFn)
 hljs.registerAliases(['ab'], { languageName: 'amber' })
@@ -71,24 +74,35 @@ class MarkdownRenderer extends Renderer {
             </div>
         `
     }
-}
 
-function initializeMarked() {
-    Marked.setBlockRule(COMPLEX_IMAGE_RULE, complexImageParser)
-    Marked.setOptions({
-        renderer: new MarkdownRenderer(),
-        breaks: true,
-        gfm: true,
-        highlight(code, lang) {
-            if (!lang) return code
-            return hljs.highlight(code, { language: lang }).value
+    link(href: string, title: string, text: string): string {
+        const assetRegex = /\/(internal|images)/
+        const version = (this.options as MarkOptions & {version: string}).version
+        if (href.startsWith('http') || assetRegex.test(window.location.pathname)) {
+            return `<a href="${href}"${title ? ` title="${title}"` : ''}>${text}</a>`;
         }
-    })
+        return `<a href="/${generateUrl(version, href.slice(1))}"${title ? ` title="${title}"` : ''}>${text}</a>`;
+    }
 }
-
-initializeMarked()
 
 export default function Markdown({ content }: { content: string }) {
+    const { version } = useVersion()
+
+    const markdownParser = useMemo(() => {
+        const renderer = new MarkdownRenderer({ ...Marked.options, version } as MarkedOptions);
+        Marked.setBlockRule(COMPLEX_IMAGE_RULE, complexImageParser)
+        Marked.setOptions({
+            renderer,
+            breaks: true,
+            gfm: true,
+            highlight(code, lang) {
+                if (!lang) return code
+                return hljs.highlight(code, { language: lang }).value
+            }
+        })
+        return Marked
+    }, [version])
+
     useEffect(() => {
         const blocks: HTMLDivElement[] = Array.from(document.querySelectorAll(`.${style.container}`))
         if (matchMedia('(hover: none)').matches) {
@@ -97,6 +111,6 @@ export default function Markdown({ content }: { content: string }) {
     }, [])
 
     return (
-        <div className={style.markdown} dangerouslySetInnerHTML={{ __html: Marked.parse(content) }} />
+        <div className={style.markdown} dangerouslySetInnerHTML={{ __html: markdownParser.parse(content) }} />
     )
 }
